requirements.txt
djangorestframework
djangorestframework_simplejwt
python-decouple
drf-yasg
drf-spectacular
apscheduler